move keyCreate & helpers into keycreate.go
