+++ fluidkeys/main.go
@@ -3,18 +3,15 @@ package main
 import (
 	"bufio"
 	"fmt"
-	"math/rand"
 	"os"
 	"path/filepath"
 	"sort"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/mitchellh/go-homedir"
 
 	"github.com/docopt/docopt-go"
-	"github.com/fluidkeys/fluidkeys/backupzip"
 	"github.com/fluidkeys/fluidkeys/colour"
 	"github.com/fluidkeys/fluidkeys/config"
 	"github.com/fluidkeys/fluidkeys/database"
@@ -26,18 +23,10 @@ import (
 	"github.com/fluidkeys/fluidkeys/out"
 	"github.com/fluidkeys/fluidkeys/pgpkey"
 	"github.com/fluidkeys/fluidkeys/scheduler"
-
-	"github.com/sethvargo/go-diceware/diceware"
 )
-
-const DicewareNumberOfWords int = 6
-const DicewareSeparator string = "."
 
 const GPGMissing string = "GPG isn't working on your system 🤒\n"
 const ContinueWithoutGPG string = "You can still use FluidKeys to make a key and then later import it from your backup.\n\nAlternatively, quit now [ctrl-c], install GPG then run FluidKeys again.\n"
-const PromptPressEnter string = "Press enter to continue"
-
-const PromptEmail string = "Enter your email address, this will help other people find your key.\n"
 
 const PromptWhichKeyFromGPG string = "Which key would you like to import?"
 
@@ -273,65 +262,6 @@ func loadPgpKeys() ([]pgpkey.PgpKey, error) {
 	return keys, nil
 }
 
-func keyCreate() exitCode {
-
-	if !gpg.IsWorking() {
-		out.Print(colour.Warning("\n" + GPGMissing + "\n"))
-		out.Print(ContinueWithoutGPG + "\n")
-		promptForInput("Press enter to continue. ")
-	}
-	out.Print("\n")
-	email := promptForEmail()
-	channel := make(chan generatePgpKeyResult)
-	go generatePgpKey(email, channel)
-
-	password := generatePassword(DicewareNumberOfWords, DicewareSeparator)
-
-	out.Print("Your key will be protected with this password:\n\n")
-	displayPassword(password)
-	if !userConfirmedRandomWord(password) {
-		out.Print("Those words did not match. Here it is again:\n\n")
-		displayPassword(password)
-		if !userConfirmedRandomWord(password) {
-			out.Print("Those words didn't match again. Quitting...\n")
-			os.Exit(1)
-		}
-	}
-
-	out.Print("Creating key for " + colour.Info(email) + ":\n\n")
-
-	generateJob := <-channel
-
-	if generateJob.err != nil {
-		panic(fmt.Sprint("Failed to generate key: ", generateJob.err))
-	}
-	printSuccessfulAction("Generate key for ian@example.com")
-
-	pushPrivateKeyBackToGpg(generateJob.pgpKey, password.AsString(), &gpg)
-	printSuccessfulAction("Store key in " + colour.Info("gpg"))
-
-	fingerprint := generateJob.pgpKey.Fingerprint()
-	db.RecordFingerprintImportedIntoGnuPG(fingerprint)
-	if err := tryEnableMaintainAutomatically(generateJob.pgpKey, password.AsString()); err == nil {
-		printSuccessfulAction("Store password in system keyring")
-		printSuccessfulAction("Setup automatic maintenance using " + colour.Info("cron"))
-	} else {
-		printFailedAction("Setup automatic maintenance")
-	}
-
-	filename, err := backupzip.OutputZipBackupFile(fluidkeysDirectory, generateJob.pgpKey, password.AsString())
-	if err != nil {
-		printFailedAction("Make a backup ZIP file")
-	}
-	directory, _ := filepath.Split(filename)
-	printSuccessfulAction("Make a backup ZIP file in")
-	out.Print("        " + colour.Info(directory) + "\n\n")
-
-	printSuccess("Successfully created key for ian@example.com")
-	out.Print("\n")
-	return 0
-}
-
 func keyList() exitCode {
 	keys, err := loadPgpKeys()
 	if err != nil {
@@ -371,12 +301,6 @@ func makeFluidkeysHomeDirectory() (string, error) {
 	fluidkeysDir := filepath.Join(homeDirectory, ".config", "fluidkeys")
 	os.MkdirAll(fluidkeysDir, 0700)
 	return fluidkeysDir, nil
-}
-
-func generatePgpKey(email string, channel chan generatePgpKeyResult) {
-	key, err := pgpkey.Generate(email, time.Now(), nil)
-
-	channel <- generatePgpKeyResult{key, err}
 }
 
 func formatListedKeysForImportingFromGpg(secretKeyListings []gpgwrapper.SecretKeyListing) string {
@@ -441,48 +365,4 @@ func promptForInputWithPipes(prompt string, reader *bufio.Reader) string {
 
 func promptForInput(prompt string) string {
 	return promptForInputWithPipes(prompt, bufio.NewReader(os.Stdin))
-}
-
-func promptForEmail() string {
-	out.Print(PromptEmail + "\n")
-	return promptForInput("[email] : ")
-}
-
-func generatePassword(numberOfWords int, separator string) DicewarePassword {
-	return DicewarePassword{
-		words:     diceware.MustGenerate(numberOfWords),
-		separator: separator,
-	}
-}
-
-func displayPassword(password DicewarePassword) {
-	out.Print("  " + colour.Info(password.AsString()) + "\n\n")
-	out.Print("If you use a password manager, save it there now.\n\n")
-	out.Print(colour.Warning("Store this safely, otherwise you won’t be able to use your key\n\n"))
-
-	promptForInput("Press enter when you've stored it safely. ")
-}
-
-func userConfirmedRandomWord(password DicewarePassword) bool {
-	clearScreen()
-	rand.Seed(time.Now().UnixNano())
-	randomIndex := rand.Intn(len(password.words))
-	correctWord := password.words[randomIndex]
-	wordOrdinal := humanize.Ordinal(randomIndex + 1)
-
-	out.Print(fmt.Sprintf("Enter the %s word from your password\n\n", wordOrdinal))
-	givenWord := promptForInput("[" + wordOrdinal + " word] : ")
-	return givenWord == correctWord
-}
-
-func clearScreen() {
-	out.Print("\033[H\033[2J")
-}
-
-// Max returns the larger of x or y.
-func Max(x, y int) int {
-	if x < y {
-		return y
-	}
-	return x
 }
